feat: 增加 ChangeDetectionCleanTask 任务类
